chore: move all binaries to bin/
Now florestad (renamed to libflorestad) and floresta-cli (now called floresta-rpc) are
library-only, and can be used by others to integrate floresta. The actual binaries
(florestad and floresta-cli) are under `bin/`, and have only the bare-minimum
for the actual binaries.

Closes #499
