+++ bin/floresta-cli/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "floresta-cli"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+description = """
+A command line interface for Florestad.
+
+You can use this client to interact with a running Florestad node. Check out doc/rpc/ or run `floresta-cli --help` for more information
+on how to use the CLI.
+"""
+repository = "https://github.com/vinteumorg/Floresta"
+readme = "README.md"
+keywords = ["bitcoin", "utreexo", "node", "blockchain", "rust"]
+categories = ["cryptography::cryptocurrencies", "command-line-utilities"]
+
+[dependencies]
+clap = { version = "4.0.29", features = ["derive"] }
+bitcoin = { version = "0.32", features = ["serde", "std"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+anyhow = "1.0"
+floresta-rpc = { path = "../../crates/floresta-rpc", features = ["clap"] }
