+++ crates/floresta-node/src/error.rs
