+++ crates/floresta-node/src/json_rpc/res.rs
