+++ crates/floresta-rpc/src/lib.rs
@@ -13,7 +13,6 @@
 #[cfg(feature = "with-jsonrpc")]
 pub mod jsonrpc_client;
 
-pub mod parsers;
 pub mod rpc;
 pub mod rpc_types;
